Implement CloudWindow and CloudList components
Description:
This commit introduces the CloudWindow and CloudList components to the project. The CloudWindow component is responsible for displaying an overlay window, while the CloudList component will render a lazy lists (i.e. list of devices ready to establish a Wi-Fi connection)

Changes Made:

- Created CloudWindow composable function
- Added CloudList composable function
- Integrated CloudList into CloudWindow-

Additional Notes:

The upcoming implementation will focus on enhancing the CloudList component to display a list of devices available for Wi-Fi connection establishment.